+++ example-vite/src/views/Dashboard.vue
@@ -1,207 +1,193 @@
 <template>
-  <div class="container">
-    <h1>A Hello Vite + Vue 3!</h1>
-    <p ref="titleRef">Edit ./App.vue to test hot module replacement (HMR).</p>
-    <p>
-      <span>Count is: {{ count }}</span>
-      <button @click="count++">increment</button>
-    </p>
-    <p><button @click="(e) => testApi('healthcheck')">Test API</button> <button @click="(e) => testApi('health-auth')">Test API Auth</button></p>
-    <p>Non-Reactive Data: {{ nonReactiveData }}</p>
-    <p>Reactive Data: {{ reactiveData }}</p>
-    <p>Vuex Store {{ storeCount }} - {{ storeUser }}</p>
-    <p>Axios GET {{ msg }}</p>
-    <p><button @click="subPn">Sub PN</button>&nbsp;<button @click="unsubPn">Unsub PN</button>&nbsp;<button @click="testPn">Test PN</button></p>
-    <ul>
-      <li v-for="n in 50" :key="n">{{ n }}</li>
-    </ul>
+  <div class="ds-stat-container">
+    <a-row :gutter="8">
+      <a-col class="ds-stat" :sm="24" :md="12" :xl="6">
+        <a-card>
+          <a-statistic
+            title="Feedback"
+            :value="11.28"
+            :precision="2"
+            suffix="%"
+            :value-style="{ color: '#3f8600' }"
+          >
+            <template #prefix>
+              <arrow-up-outlined />
+            </template>
+          </a-statistic>
+        </a-card>
+      </a-col>
+      <a-col class="ds-stat" :sm="24" :md="12" :xl="6">
+        <a-card>
+          <a-statistic
+            title="Idle"
+            :value="9.3"
+            :precision="2"
+            suffix="%"
+            class="demo-class"
+            :value-style="{ color: '#cf1322' }"
+          >
+            <template #prefix>
+              <arrow-down-outlined />
+            </template>
+          </a-statistic>
+        </a-card>
+      </a-col>
+
+      <a-col class="ds-stat" :sm="24" :md="12" :xl="6">
+        <a-card>
+          <a-statistic title="Active Users" :value="112893" />
+        </a-card>
+      </a-col>
+      <a-col class="ds-stat" :sm="24" :md="12" :xl="6">
+        <a-card>
+          <a-statistic title="Account Balance (CNY)" :precision="2" :value="112893" />
+        </a-card>
+      </a-col>
+    </a-row>
+
+    <a-row :gutter="8">
+      <a-col class="ds-stat" :lg="24" :xl="16">
+
+      <a-row :gutter="8">
+        <a-col class="ds-stat" :span="12">
+          <a-page-header title="Team Members" />
+          <a-list :grid="{ gutter: 16, xs: 1, sm: 1, md: 2, lg: 3, xl: 3, xxl: 3 }" :data-source="data">
+            <template #renderItem="{ item }">
+              <a-list-item>
+                <a-card :bordered="false">
+                  <a-avatar :size="64">
+                    <template #icon><UserOutlined /></template>
+                  </a-avatar>
+                  <a-card-meta>
+                    <template #title><span>{{ item.name }}</span></template>
+                    <template #description><span style="font-size: 0.8em;">{{ item.title }}</span></template>
+                  </a-card-meta>
+                </a-card>
+              </a-list-item>
+            </template>
+          </a-list>
+        </a-col>
+        <a-col class="ds-stat" :span="12">
+          <a-page-header title="Comments" />
+          <a-list item-layout="horizontal" :data-source="data">
+            <template #renderItem="{ item }">
+              <a-list-item>
+                <a-list-item-meta
+                  description="Ant Design, a design language for background applications, is refined by Ant UED Team"
+                >
+                  <template #title>
+                    <a href="https://www.antdv.com/">{{ item.title }}</a>
+                  </template>
+                  <template #avatar>
+                    <a-avatar src="https://zos.alipayobjects.com/rmsportal/ODTLcjxAfvqbxHnVXCYX.png" />
+                  </template>
+                </a-list-item-meta>
+              </a-list-item>
+            </template>
+          </a-list>
+        </a-col>
+
+        <a-col class="ds-stat" :span="24">
+          <a-page-header title="Faked Quality" />
+          <MultiChart />
+        </a-col>
+        <a-col class="ds-stat" :span="24">
+        </a-col>
+      </a-row>
+
+      </a-col>
+
+      <a-col class="ds-stat" :lg="24" :xl="8">
+        <a-row :gutter="8">
+          <a-col class="ds-stat" :span="24">
+            <a-page-header title="Pie Quality" sub-title="Some other random description here. Does not have to be too long a text..." />
+            <PieChart />
+          </a-col>
+          <a-col class="ds-stat" :span="24">
+            <a-page-header title="Useful Links" />
+            <a-skeleton :paragraph="{ rows: 3 }" />
+          </a-col>
+        </a-row>
+      </a-col>
+    </a-row>
   </div>
+
+<!--
+  <a-row :gutter="8">
+    <a-col class="ds-stat" :xs="24" :sm="12" :md="8" :lg="6">
+      <a-statistic title="Feedback" :value="1128" style="margin-right: 50px">
+        <template #suffix><like-outlined /></template>
+      </a-statistic>
+    </a-col>
+    <a-col class="ds-stat" :xs="24" :sm="12" :md="8" :lg="6">
+      <a-statistic title="Unmerged" :value="93" class="demo-class">
+        <template #suffix><span>/ 100</span></template>
+      </a-statistic>
+    </a-col>
+  </a-row>
+-->
 </template>
 
 <script>
-// defineComponent, getCurrentInstance, reactive, readonly, watch, watchEffect
-// provide, inject ???
-import { onMounted, onUpdated, onUnmounted, onBeforeUnmount, ref, computed, inject } from 'vue'
-import { useStore } from 'vuex'
-// import { useRouter, useRoute } from 'vue-router'
-import { debounce } from '../../../@es-labs/esm/util.js' // served from express /esm static route
-import { webpushSubscribe, webpushUnsubscribe, fcmSubscribe } from '../../../@es-labs/esm/pwa.js' // served from express /esm static route
-import { useXhr } from '/src/plugins/xhr.js'
-import { VITE_PWA_PN } from '/config.js'
+import { LikeOutlined, ArrowUpOutlined, ArrowDownOutlined, UserOutlined } from '@ant-design/icons-vue'
+import MultiChart from '../components/Dashboard/MultiChart.vue'
+import PieChart from '../components/Dashboard/PieChart.vue'
+
+const data = [
+  {
+    name: 'Aaron',
+    title: 'Fullstack Dev',
+  },
+  {
+    name: 'Jermyn',
+    title: 'Data Scientist',
+  },
+  {
+    name: 'Valerie',
+    title: 'Data Scientist',
+  },
+  {
+    name: 'Heping',
+    title: 'Data Scientist',
+  },
+];
 
 export default {
-  name: 'Dashboard',
+  components: {
+    LikeOutlined,
+    ArrowUpOutlined,
+    ArrowDownOutlined,
+    UserOutlined,
 
-  setup(props, context) {
-    console.log('provide-inject MyTheme', inject('MyTheme'))
-    const http = useXhr()
-    // reactivity
-    // // in provider
-    // const themeRef = ref('dark')
-    // provide(SubThemeSymbol, themeRef)
-    // // in consumer
-    // const theme = inject(SubThemeSymbol, ref('light'))
-    // watchEffect(() => console.log(`theme set to: ${theme.value}`))
-
-    const store = useStore()
-    // const route = useRoute()
-    // const router = useRouter()
-
-    // const obj = reactive({ count: 0 })
-    const count = ref(0)
-    const msg = ref('')
-    const titleRef = ref(null)
-    let nonReactiveData = 10
-    const reactiveData = ref(20)
-
-    // const plusOne = computed(() => count.value + 1)
-    const storeCount = computed(() => store.state.count) // ctx.root.$store.myModule.state.blabla
-    const storeUser = computed(() => store.state.user)
-
-    // const stop = watchEffect(() => console.log(count.value))
-    // // -> logs 0
-    // setTimeout(() => {
-    //   count.value++
-    //   // -> logs 1
-    // }, 100)
-    // // stop()
-
-    // put watchEffect inside onMounted to have access to DOM
-
-    // // watching a getter
-    // const state = reactive({ count: 0 })
-    // watch(
-    //   () => state.count,
-    //   (count, prevCount) => {
-    //   }
-    // )
-
-    // // directly watching a ref
-    // const count = ref(0)
-    // watch(count, (count, prevCount) => {
-    // })
-
-    // watch([fooRef, barRef], ([foo, bar], [prevFoo, prevBar]) => {
-    // })
-
-    // watch(
-    //   () => object_or_primitive_being_watched,
-    //   (state, prevState) => {
-    //     console.log(
-    //       "deep ",
-    //       state.attributes.name,
-    //       prevState.attributes.name
-    //     );
-    //   },
-    //   { deep: true }
-    // )
-    // watchEffect ... ?
-
-    let timerId
-    onMounted(async () => {
-      console.log('dashboard mounted!')
-      // console.log('props', props)
-      // console.log('context', context)
-      // console.log('useStore', store)
-      // console.log('useRouter', router)
-      // console.log('useRoute', route)
-      // console.log("template ref titleRef", titleRef.value)
-
-      timerId = setInterval(() => {
-        console.log('timer fired', String(selected.value))
-        nonReactiveData += 1
-        reactiveData.value += 1
-      }, 200000)
-    })
-    onBeforeUnmount(() => {
-      if (timerId) clearInterval(timerId)
-      // / console.log('dash before unmount!')
-    })
-    onUpdated(() => console.log('dash updated!'))
-    onUnmounted(() => console.log('dashboard unmounted!'))
-
-    const testApi = async (test) => {
-      try {
-        const { data } = await http.get('/api/' + test)
-        console.log('testApi', data)
-      } catch (e) {
-        console.log('testApi err', e)
-      }
-    }
-
-    const subPn = async () => {
-      console.log(VITE_PWA_PN)
-      try {
-        let subscription
-        if (VITE_PWA_PN === 'FCM') {
-          subscription = await fcmSubscribe(window.SW_REG, async (token) => {
-            await http.post('/api/webpush/sub', { subscription: token })
-          })
-        } else if (VITE_PWA_PN === 'Webpush') {
-          const { data } = await http.get('/api/webpush/vapid-public-key')
-          subscription = await webpushSubscribe(data.publicKey)
-        }
-        if (subscription) await http.post('/api/webpush/sub', { subscription })
-      } catch (e) {
-        console.log('subPn', e)
-      }
-    }
-
-    const unsubPn = async () => {
-      // No FCM Unsub
-      try {
-        if (VITE_PWA_PN === 'Webpush') await webpushUnsubscribe()
-        await http.post('/api/webpush/unsub')
-      } catch (e) {
-        console.log('unsubPn', e)
-      }
-    }
-
-    const testPn = async () => {
-      try {
-        let data
-        console.log('testPn VITE_PWA_PN', VITE_PWA_PN)
-        if (VITE_PWA_PN === 'FCM') data = { title: 'Hello', body: new Date().toLocaleString() }
-        else if (VITE_PWA_PN === 'Webpush') data = 'Hello ' + new Date().toLocaleString()
-        console.log('testPn data', data)
-        await http.post('/api/webpush/send/1', { mode: VITE_PWA_PN, data })
-      } catch (e) {
-        console.log('testPn', e)
-      }
-    }
-
-    // // Watch prop value change and assign to value 'selected' Ref
-    // watch(() => props.value, (newValue: Props['value']) => {
-    //   selected.value = newValue;
-    // });
+    PieChart,
+    MultiChart
+  },
+  setup() {
     return {
-      nonReactiveData, // non reactive
-      reactiveData, // ref reactive
-      count, // ref
-      msg,
-      titleRef,
-      storeCount, // store
-      storeUser,
-      testApi, // test API
-      subPn, // push notifications
-      unsubPn,
-      testPn
+      data,
     }
   }
 }
 </script>
 
-<style scoped>
-h1 {
-  color: #4fc08d;
+<style>
+.ds-stat {
+  margin-top: 0px;
+  padding-top: 0px;
+  /* background-color: lightslategray; */
 }
 
-h1,
-p {
-  font-family: Arial, Helvetica, sans-serif;
+.ds-stat-container {
+  /* background: #ececec; */
+  padding: 0;
 }
 
-vcxwc-sign-pad {
-  --vcxwc-sign-pad-background-color: #faa;
+.ant-card-meta-title {
+  margin-bottom: 0 !important;
+}
+
+.ant-card-body {
+  /* justify-content: center !important; */
+  text-align: center;
 }
 </style>