+++ example-vite/src/views/demo/DemoWebCam.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script>
-import '../../../@es-labs/esm/web-cam.js' // served from express /esm static route
+import '/@es-labs/esm/web-cam.js' // served from express /esm static route
 
 export default {
   name: 'WebCam',